fix: resolve all clippy lints newly visible under --all-targets
Adding --all-targets to pre-commit clippy exposed several latent lints:
- assertions_on_constants in hologram-archive section test → const { assert! }
- approx_constant in calculator example → std::f64::consts::PI
- useless_vec in hologram-ffi exec tests → array literals [42u8], [100u8]
- unnecessary_cast + unused_parens in hologram-bench ffi bench → i % 21

Also syncs pre-commit clippy to use --all-targets to match CI.

+++ crates/hologram-bench/benches/ffi.rs
@@ -15,13 +15,7 @@ fn bench_ffi_graph_build(c: &mut Criterion) {
             hologram_graph_builder_node_from_input(builder, 0, 0, 0);
             for i in 0..8 {
                 let inputs = [i as usize];
-                hologram_graph_builder_node_with_inputs(
-                    builder,
-                    3,
-                    (i % 21) as i32,
-                    inputs.as_ptr(),
-                    1,
-                );
+                hologram_graph_builder_node_with_inputs(builder, 3, i % 21, inputs.as_ptr(), 1);
             }
             let inputs = [8usize];
             hologram_graph_builder_node_with_inputs(builder, 1, 0, inputs.as_ptr(), 1);
